added python scripting to get data, then parse it  in the dom later, and display as graph
