+++ data-collection.py
@@ -1,17 +1,24 @@
 import psycopg2
 
-conn = psycopg2.connect(
-    dbname="my_local_db",
-    user="postgres",  
-    password="pump",   
-    host="localhost",
-    port="5432"
-)
+import json
 
-cur = conn.cursor()
+datafromdb = [1, 2, 3, 4, 5]
 
-cur.execute("SELECT * FROM users;")
+with open('public/data.json', 'w') as f:
+    json.dump(datafromdb, f)
 
-rows = cur.fetchall()
-for row in rows:
-    print(row)
+# conn = psycopg2.connect(
+#     dbname="my_local_db",
+#     user="postgres",  
+#     password="pump",   
+#     host="localhost",
+#     port="5432"
+# )
+
+# cur = conn.cursor()
+
+# cur.execute("SELECT * FROM users;")
+
+# rows = cur.fetchall()
+# for row in rows:
+#     print(row)
