+++ src/App.jsx
@@ -1,38 +1,32 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useState, useRef } from 'react';
 import * as d3 from 'd3';
+import generateGraph from './api/MockData';
 
 function App() {
+
   const chartRef = useRef(null);
+  const [datainput, setDatainput] = useState(null); // ✅ tracked state
 
+  // Load JSON data
   useEffect(() => {
-    const data = [30, 80, 45, 60, 20, 90, 50];
-    const svg = d3.select(chartRef.current);
-    const width = 600;
-    const height = 400;
-    const barWidth = width / data.length;
-
-    svg.selectAll("rect").remove(); // Prevent duplication
-
-    svg
-      .attr("width", width)
-      .attr("height", height)
-      .selectAll("rect")
-      .data(data)
-      .enter()
-      .append("rect")
-      .attr("x", (_, i) => i * barWidth)
-      .attr("y", d => height - d * 4)
-      .attr("width", barWidth - 5)
-      .attr("height", d => d * 4)
-      .attr("fill", "steelblue");
+    fetch('/data.json')
+      .then(res => res.json())
+      .then(data => {
+        setDatainput(data); // ✅ store it reactively
+      });
   }, []);
+
+  // Render chart after data loads
+  useEffect(() => {
+    if (datainput) {
+      generateGraph(chartRef, datainput); // ✅ safe to call here
+    }
+  }, [datainput]);
 
   return (
     <div>
       <h1>My SCADA Graph</h1>
-      <svg ref={chartRef}></svg>
-      <p>Lower Reservoir</p>
-      <p>Upper Reservoir</p>
+      <svg ref = {chartRef}></svg>
     </div>
   );
 }
